% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcint.R
\name{calc_mcint}
\alias{calc_mcint}
\title{Calculate Monte Carlo Integral}
\usage{
calc_mcint(N, x_l = 0, x_u = 1, func = function(x) x^2)
}
\arguments{
\item{N}{the number of random points}

\item{x_l, }{x_u the lower and upper bounds on x, respectively}

\item{func}{the function whose integral we are estimating}
}
\value{
Estimated integral given by Monte Carlo integration.
}
\description{
Takes in any numeric value and squares it.
}
