% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_under.R
\name{is_under}
\alias{is_under}
\title{Calculates if a point is under a curve}
\usage{
is_under(func, x, y)
}
\arguments{
\item{func}{the function of interest}

\item{x, }{y the coordinates of the point of interest}
}
\value{
The boolean (x, y) <= (x, func(x))
}
\description{
The point (x, y) is under a curve func(x) if y <= func(x)
}
